add review.Authenticate Base

+++ reviews/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render, redirect
 from .models import Review, Comment
 from .forms import ReviewForm, CommentForm
+from django.contrib.auth.decorators import login_required
 
 # Create your views here.
 
@@ -11,7 +12,7 @@ def index(request):
     }
     return render(request, 'reviews/index.html', context)
 
-
+@login_required
 def create(request):
     if request.method == "POST":
         form = ReviewForm(request.POST, request.FILES)
@@ -39,26 +40,32 @@ def detail(request, review_pk):
     }
     return render(request, 'reviews/detail.html', context)
 
+@login_required
 def delete(request, review_pk):
     review = Review.objects.get(pk=review_pk)
-    review.delete()
+    if request.user == review.user:
+        review.delete()
     return redirect('reviews:index')
 
+@login_required
 def update(request, review_pk):
     review = Review.objects.get(pk=review_pk)
-    if request.method == 'POST':
-        form = ReviewForm(request.POST, request.FILES, instance=review)
-        if form.is_valid():
-            form.save()
-            return redirect('reviews:detail', review_pk)
+    if request.user == review.user:
+        if request.method == 'POST':
+            form = ReviewForm(request.POST, request.FILES, instance=review)
+            if form.is_valid():
+                form.save()
+                return redirect('reviews:detail', review_pk)
+        else:
+            form = ReviewForm(instance=review)
+        context = {
+            'form':form
+        }
+        return render(request, 'reviews/update.html', context)
     else:
-        form = ReviewForm(instance=review)
-    context = {
-        'form':form
-    }
-    return render(request, 'reviews/update.html', context)
+        return redirect('reviews:detail', review.pk)
 
-
+@login_required
 def comment_create(request, review_pk):
     review = Review.objects.get(pk=review_pk)
     comment_form = CommentForm(request.POST)
@@ -74,8 +81,9 @@ def comment_create(request, review_pk):
     }
     return render(request, 'reviews/detail.html', context)
 
-
+@login_required
 def comment_delete(request, review_pk, comment_pk):
     comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
+    if request.user == comment.user:
+        comment.delete()
     return redirect('reviews:detail', review_pk)
